Improve README instructions
I notice in #22 Gemfile and Gemfile.lock were added to make sure contributors all use the right version of Jekyll. The problem is that the commands given in the README don't actually enforce this at all, because they're run outside of Bundler.

This pull requests changes the instructions in the README to actually use Bundler so the Jekyll verison used will always be correct as was the intention in #22.